view model koin injection released
