+++ app/build.gradle
@@ -47,6 +47,8 @@ dependencies {
 
     implementation 'io.coil-kt:coil:1.4.0'
 
+    implementation "io.insert-koin:koin-android:3.1.6"
+
     implementation "androidx.fragment:fragment-ktx:1.4.1"
     implementation 'androidx.core:core-ktx:1.7.0'
     implementation 'androidx.appcompat:appcompat:1.4.1'
