+++ app/src/main/java/com/ineedyourcode/githubapiapp/domain/repository/UsecaseRepository.kt
@@ -4,5 +4,4 @@ import com.ineedyourcode.githubapiapp.domain.usecase.GetProjectRepositoryUsecase
 import com.ineedyourcode.githubapiapp.domain.usecase.GetUserUsecase
 import com.ineedyourcode.githubapiapp.domain.usecase.SearchUserUsecase
 
-interface UsecaseRepository : SearchUserUsecase, GetUserUsecase, GetProjectRepositoryUsecase {
-}
+interface UsecaseRepository : SearchUserUsecase, GetUserUsecase, GetProjectRepositoryUsecase
